Beginnings of login functionality
The login form will self submit user input and check that input against the database. If a result is found, the user information is stored to a session. Currently there is nothing stopping multiple users from logging in, and the login page does not advance to another webpage. Additional pages could be similar that re designed to deal with a logged in user