Fixed the login functionality
Separated the login code, the HTML was brought into the index.html file to be an SPA. The PHP script was put into a php folder called login.php. login.js was also created to keep the JS code separate as well.
The login and logout navigation tags/pages are hidden to the client depending on login status. Logging in redirects back to the home page