Differentiation between staff and user logins
If the user is a staff member, they will have access to the staff part of the webpage, if not, the user will not be able to see the staff page and get access to it.
Fixes have been made to each JS file and PHP file. The paths should be correct and the SQL should be correct also for usage on CETO