Implemented shopping cart and company info
users (registered and unregistered) can now add products to their shopping cart and checkout. Products are removed from the database on checkout. Shopping cart is unfortunately not maintained through login i.e. if a user adds products to their shopping cart and then logs in, their shopping cart will be reset. The actual shopping cart is not deleted; it still exists in the SQL database, the user just doesn't have access to it. When a user adds a product to their cart, a query is sent deleting all the cart items more than a week old. This means all shopping carts are cleared after a week.

Added company information to home page and footer.

Generally improved styling.